calculate.overlap {VennDiagram}	R Documentation

Calculate Overlap

Description

Determine the groupings of values as they would be presented in the venn diagram.

Usage

calculate.overlap(x)

Arguments

x
A list of vectors (e.g., integers, chars), with each component corresponding to a separate circle in the Venn diagram

Details

This function mostly complements the venn.diagram() function for the case where users want to know what values are grouped into the particular areas of the venn diagram.

Value

Returns a list of lists which contain the values assigned to each of the areas of a venn diagram.

Author(s)

Christopher Lalansingh

Examples

# A simple single-set diagram
cardiome <- letters[1:10]
superset <- letters[8:24]
overlap <- calculate.overlap(
	x = list(
		"Cardiome" = cardiome,
		"SuperSet" = superset
		)
	);

[Package VennDiagram version 1.6.20 Index]